+++ src/main/java/com/example/walkingmate_back/user/dto/UserBodyResponseDTO.java
@@ -10,14 +10,14 @@ public class UserBodyResponseDTO {
 
     private String userId;  // 사용자 id
 
-    private int height;  // 사용자 키
+    private double height;  // 사용자 키
 
-    private int weight;  // 사용자 몸무게
+    private double weight;  // 사용자 몸무게
 
-    private int BMI;  // BMI
+    private double BMI;  // BMI
 
     @Builder
-    public UserBodyResponseDTO(String userId, int weight, int height, int BMI) {
+    public UserBodyResponseDTO(String userId, double weight, double height, double BMI) {
         this.userId=userId;
         this.weight=weight;
         this.height=height;
